services/worker/requirements.txt
certifi==2025.10.5
clickhouse-connect==0.9.2
lz4==4.4.4
nats-py==2.12.0
orjson==3.11.4
psycopg==3.2.12
psycopg-binary==3.2.12
pytz==2025.2
typing_extensions==4.15.0
urllib3==2.5.0
zstandard==0.25.0
